Add crowd cheering
