+++ Assets/Models/Characters/Animation/Clapping.anim
@@ -19327,7 +19327,7 @@ AnimationClip:
     intParameter: 0
     messageOptions: 0
   - time: 0.3
-    functionName: 
+    functionName: StartClap
     data: 
     objectReferenceParameter: {fileID: 0}
     floatParameter: 0
