+++ Assets/Plugins/FMOD/Cache/Editor/FMODStudioCache.asset
@@ -676,33 +676,6 @@ MonoBehaviour:
   MinDistance: 1
   MaxDistance: 20
   Length: 500
---- !u!114 &-4901362813040367991
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: ceb653cd98e289a4e8697a1af55201f2, type: 3}
-  m_Name: event:/Audience/Wave
-  m_EditorClassIdentifier: 
-  Path: event:/Audience/Wave
-  Guid:
-    Data1: 750302676
-    Data2: 1156063563
-    Data3: -1326784345
-    Data4: -458266410
-  Banks:
-  - {fileID: 1278483287379617233}
-  IsStream: 0
-  Is3D: 1
-  IsOneShot: 1
-  Parameters: []
-  MinDistance: 1
-  MaxDistance: 20
-  Length: 6000
 --- !u!114 &-4686498582662505318
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -1066,6 +1039,33 @@ MonoBehaviour:
   MinDistance: 1
   MaxDistance: 20
   Length: 900
+--- !u!114 &-3512891166489636755
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: ceb653cd98e289a4e8697a1af55201f2, type: 3}
+  m_Name: event:/Audience/Cheering
+  m_EditorClassIdentifier: 
+  Path: event:/Audience/Cheering
+  Guid:
+    Data1: -368239931
+    Data2: 1178626115
+    Data3: -706465407
+    Data4: -2100007398
+  Banks:
+  - {fileID: 1278483287379617233}
+  IsStream: 0
+  Is3D: 1
+  IsOneShot: 1
+  Parameters: []
+  MinDistance: 1
+  MaxDistance: 20
+  Length: 3683
 --- !u!114 &-3472946857483378134
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -1455,6 +1455,33 @@ MonoBehaviour:
   MinDistance: 1
   MaxDistance: 20
   Length: 5351
+--- !u!114 &-949703858989825250
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: ceb653cd98e289a4e8697a1af55201f2, type: 3}
+  m_Name: event:/Audience/Waving
+  m_EditorClassIdentifier: 
+  Path: event:/Audience/Waving
+  Guid:
+    Data1: 750302676
+    Data2: 1156063563
+    Data3: -1326784345
+    Data4: -458266410
+  Banks:
+  - {fileID: 1278483287379617233}
+  IsStream: 0
+  Is3D: 1
+  IsOneShot: 1
+  Parameters: []
+  MinDistance: 1
+  MaxDistance: 20
+  Length: 6000
 --- !u!114 &-671411267369357733
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -1483,33 +1510,6 @@ MonoBehaviour:
   MinDistance: 1
   MaxDistance: 20
   Length: 500
---- !u!114 &-553678846387809314
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: ceb653cd98e289a4e8697a1af55201f2, type: 3}
-  m_Name: event:/Audience/Whistling
-  m_EditorClassIdentifier: 
-  Path: event:/Audience/Whistling
-  Guid:
-    Data1: -1179826819
-    Data2: 1133699828
-    Data3: -495157333
-    Data4: -1497646473
-  Banks:
-  - {fileID: 1278483287379617233}
-  IsStream: 0
-  Is3D: 1
-  IsOneShot: 1
-  Parameters: []
-  MinDistance: 1
-  MaxDistance: 20
-  Length: 1800
 --- !u!114 &-209184966175457103
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -1555,11 +1555,9 @@ MonoBehaviour:
   - {fileID: 1278483287379617233}
   EditorEvents:
   - {fileID: -6996258878270670034}
-  - {fileID: -553678846387809314}
   - {fileID: 4950022710509616201}
   - {fileID: -3385780560051107157}
   - {fileID: -2058432523618543073}
-  - {fileID: -4901362813040367991}
   - {fileID: -6919760833295164659}
   - {fileID: 937579986535423678}
   - {fileID: 7602683481376795458}
@@ -1612,12 +1610,14 @@ MonoBehaviour:
   - {fileID: -1246386247636295271}
   - {fileID: -3556818849776593372}
   - {fileID: 3424449672764231202}
+  - {fileID: -3512891166489636755}
+  - {fileID: -949703858989825250}
   EditorParameters: []
   MasterBanks:
   - {fileID: 1278483287379617233}
   StringsBanks:
   - {fileID: 3916383380135203157}
-  cacheTime: 638014604303259869
+  cacheTime: 638016140705983358
   cacheVersion: 131591
 --- !u!114 &12281985263074687
 MonoBehaviour:
@@ -1902,10 +1902,10 @@ MonoBehaviour:
   Path: RoomForSound_FMOD/Build/Desktop/Master.bank
   Name: Master
   StudioPath: bank:/Master
-  lastModified: 638014604303239871
+  lastModified: 638016140705925173
   FileSizes:
   - Name: Desktop
-    Value: 8489696
+    Value: 8491680
   Exists: 1
 --- !u!114 &1281051470542862836
 MonoBehaviour:
@@ -2281,12 +2281,12 @@ MonoBehaviour:
   Path: RoomForSound_FMOD/Build/Desktop/Master.strings.bank
   Name: Master.strings
   StudioPath: bank:/Master.strings
-  lastModified: 638014604303259869
+  lastModified: 638016140705983358
   FileSizes:
   - Name: Desktop
-    Value: 3284
+    Value: 3286
   - Name: Mobile
-    Value: 3284
+    Value: 3286
   Exists: 1
 --- !u!114 &4099776152855883723
 MonoBehaviour:
