+++ Assets/Scripts/Music Elements/AudienceAudio.cs
@@ -1,5 +1,3 @@
-using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
 
 public class AudienceAudio : MonoBehaviour
@@ -12,6 +10,14 @@ public class AudienceAudio : MonoBehaviour
     public void StartClap()
     {
         clap = FMODUnity.RuntimeManager.CreateInstance("event:/Audience/Clapping");
+        clap.set3DAttributes(FMODUnity.RuntimeUtils.To3DAttributes(gameObject));
+        clap.start();
+        clap.release();
+    }
+
+    public void StartCheering()
+    {
+        clap = FMODUnity.RuntimeManager.CreateInstance("event:/Audience/Cheering");
         clap.set3DAttributes(FMODUnity.RuntimeUtils.To3DAttributes(gameObject));
         clap.start();
         clap.release();
