--- Assets/Scripts/Music Elements/AudioMarkerManager copy 2.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 552fb396812b264479e049ca7061feba
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
