--- Assets/Scripts/Music Elements/AudioMarkerManager copy.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f4237fd8d99dd2e428afaad7e848e5c5
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
