--- Assets/Scripts/Music Elements/AudioMarkerManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8743a85cf59959d4382eae6b5d179107
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
