--- Assets/Scripts/StandNoteControl.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 85442987259a3754a9833401c992a65c
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
