--- Assets/Scripts/XylophoneKey.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: cb21a255ee9c5eb41a694027a8a75f66
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
